requirements.txt
fastapi==0.110.2
pandas==2.1.3
google-cloud-storage==2.16.0
google-api-python-client==2.126.0
google-auth==2.29.0
uvicorn==0.29.0
requests==2.31.0
isodate==0.6.1